Add caching + refactor network / graph visualization

- flag publications with invalid documentation
- perform simple url classification 
- ping listed code archive urls using the requests library to determine if the url is still valid.

+++ citation/serializers.py
@@ -206,7 +206,7 @@ class PublicationAuditCommandSerializer(serializers.Serializer):
 
 class AuditCommandContibutionSerializer(serializers.Serializer):
     creator = serializers.StringRelatedField(read_only=True)
-    contribution = serializers.IntegerField(read_only=True)
+    contribution = serializers.FloatField(read_only=True)
     date_added = serializers.DateTimeField(read_only=True, format='%m/%d/%Y %H:%M')
